+++ source/Samples/ImageReconstructionSolution/ImageReconstructionSolution.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35527.113 d17.12
+VisualStudioVersion = 17.12.35527.113
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NeoCortexApiSample1.Tests", "..\NeoCortexApiSample1.Tests\NeoCortexApiSample1.Tests.csproj", "{1F3BE2E2-8854-470A-AF06-83FC4F44B432}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NeoCortexApiSample", "..\NeoCortexApiSample\NeoCortexApiSample.csproj", "{43F1D754-0D58-4418-8179-59813C13715F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NeoCortexApiSample1.Tests", "..\..\UnitTestsProject\NeoCortexApiSample1.Tests\NeoCortexApiSample1.Tests.csproj", "{3DB48DCF-A3D9-4537-834E-28BD8DB81A3E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1F3BE2E2-8854-470A-AF06-83FC4F44B432}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1F3BE2E2-8854-470A-AF06-83FC4F44B432}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1F3BE2E2-8854-470A-AF06-83FC4F44B432}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1F3BE2E2-8854-470A-AF06-83FC4F44B432}.Release|Any CPU.Build.0 = Release|Any CPU
 		{43F1D754-0D58-4418-8179-59813C13715F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{43F1D754-0D58-4418-8179-59813C13715F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{43F1D754-0D58-4418-8179-59813C13715F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{43F1D754-0D58-4418-8179-59813C13715F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3DB48DCF-A3D9-4537-834E-28BD8DB81A3E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3DB48DCF-A3D9-4537-834E-28BD8DB81A3E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3DB48DCF-A3D9-4537-834E-28BD8DB81A3E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3DB48DCF-A3D9-4537-834E-28BD8DB81A3E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
