--- source/Samples/NeoCortexApiSample/ImageReconstructorWithClassifiers.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-//False one for testing binarization
-//comment for testing
-namespace NeoCortexApiSample
-{
-    //testing dates
-    internal class ImageReconstructorWithClassifiers
-    {
-
-    }
-}
