+++ source/UnitTestsProject/NeoCortexApiSample1.Tests/TestImages/Car.png
