+++ source/UnitTestsProject/NeoCortexApiSample1.Tests/TestImages/flower.png
